feat(load-balancers): add monitor-groups/index, monitor-groups/monitor-groups, monitor-groups/references
